add support for "\s" and "\d"; squash bugs
